Yes_no_value_report generator add (to align with the other generators)
